backend/requirements.txt
# FastAPI & Web Framework
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
websockets==12.0

# Database
sqlalchemy==2.0.32
alembic==1.13.2
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Redis & Celery
redis==5.0.8
celery==5.4.0
flower==2.0.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
pydantic==2.9.0
pydantic-settings==2.5.0
email-validator==2.1.0
slowapi==0.1.9

# AI & NLP
openai==1.43.0
langchain==0.3.0
langchain-community==0.3.0
langchain-text-splitters==0.3.0
langgraph==0.2.45
llama-index==0.11.0
llama-index-vector-stores-qdrant==0.3.0
qdrant-client==1.11.0
neo4j==5.26.0
chromadb==0.5.5

# Embeddings & NLP
sentence-transformers==3.0.1
spacy==3.7.5
tiktoken==0.7.0

# Data Processing
numpy==1.26.4
pandas==2.2.2

# HTTP Client
httpx==0.27.2
aiohttp==3.10.5

# File Processing
pypdf==4.3.1
python-docx==1.1.2
ebooklib==0.18

# Monitoring & Logging
sentry-sdk[fastapi]==2.14.0

# Testing
pytest==8.3.2
pytest-asyncio==0.24.0
pytest-cov==5.0.0
httpx==0.27.2

# Utils
python-slugify==8.0.4
pydantic-core==2.23.2
